+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Global bonuses did not always sum up correctly, e.g. Aura of Protection and Cloak of Protection could add up to 51, not 6.
 * Some spell icon improvements from @Raccoomph#8480
+* Hand of Healing was not parsing as Heal action.
 
 # 3.3.4